docs: document game HUD and layers
